move some more utility functions from tilesApi to utils
