+++ GATEWAY/src/providers/tilesApi.service.ts
@@ -30,29 +30,6 @@ export class TilesApi {
   };
 
   /** 
-   * Returns an object with name and properties from the inputstring
-   * @param {string} eventString - A string on the format eventName,properties...
-   */
-  getEventStringAsObject = (eventString: string): CommandObject => {
-    const params = eventString.split(',');
-    if (params.length > 1){
-      return {
-          name: params[0],
-          properties: Array.prototype.slice.call(params, 1)
-      };
-    }
-    return null;
-  };
-
-  /** 
-   * Returns a string from the given commandObject
-   * @param {CommansObject} cmdObj - the command to turn into a string
-   */
-  getCommandObjectAsString = (cmdObj: CommandObject): string => {
-    return `${cmdObj.name},${cmdObj.properties.toString()}`;
-  };
-
-  /** 
    * Tests if a device is a tile
    * @param {any} device - the device to test
    */
